Kernel/FATFS: Abstract out SFN helpers and make them usable in userspace
